+++ lib/pingpp/balance_settlement.rb
@@ -1,0 +1,9 @@
+module Pingpp
+  class BalanceSettlement < AppBasedResource
+    extend Pingpp::APIOperations::List
+
+    def self.object_name
+      'balance_settlement'
+    end
+  end
+end